requirements.txt
asyncclick==8.1.3.4
anyio==3.6.2
nginxparser-eb==0.0.9
boto3==1.26.140
grpcio==1.54.2
google==3.0.0
protobuf==4.23.1
alive-progress==3.1.2